docs: ✏️ update Autoformer arch

+++ basicts/archs/arch_zoo/utils/xformer.py
@@ -36,4 +36,4 @@ def data_transformation_4_xformer(history_data: torch.Tensor, future_data: torch
         x_mark_dec_future = future_data[..., :, 0, 1:]          # B, L2, C-1
         x_mark_dec = torch.cat([x_mark_dec_token, x_mark_dec_future], dim=1)    # B, (start_token_length+L2), C-1
 
-    return x_enc, x_mark_enc, x_dec, x_mark_dec
+    return x_enc.float(), x_mark_enc.float(), x_dec.float(), x_mark_dec.float()
